Move execute() to its own API namespace, disambiguate (#580)
