+++ docs/sections/user_guide/api/execute.rst
@@ -1,0 +1,6 @@
+``uwtools.api.execute``
+=======================
+
+.. automodule:: uwtools.api.execute
+   :inherited-members:
+   :members:
