+++ docs/sections/user_guide/api/index.rst
@@ -7,6 +7,7 @@ API
    config
    driver
    esg_grid
+   execute
    file
    filter_topo
    fv3
