+++ src/uwtools/scheduler.py
@@ -14,7 +14,7 @@ from typing import Any, Optional
 from uwtools.exceptions import UWConfigError
 from uwtools.logging import log
 from uwtools.strings import STR
-from uwtools.utils.processing import execute
+from uwtools.utils.processing import run_shell_cmd
 
 
 class JobScheduler(ABC):
@@ -80,7 +80,7 @@ class JobScheduler(ABC):
         cmd = f"{self._submit_cmd} {runscript}"
         if submit_file:
             cmd += " 2>&1 | tee %s" % submit_file
-        success, _ = execute(cmd=cmd, cwd=f"{runscript.parent}")
+        success, _ = run_shell_cmd(cmd=cmd, cwd=f"{runscript.parent}")
         return success
 
     # Private methods
