+++ src/uwtools/tests/drivers/test_driver.py
@@ -376,9 +376,9 @@ def test_Driver__run_via_local_execution(driverobj):
     executable = Path(driverobj.config["execution"]["executable"])
     executable.touch()
     with patch.object(driverobj, "provisioned_rundir") as prd:
-        with patch.object(driver, "execute") as execute:
+        with patch.object(driver, "run_shell_cmd") as run_shell_cmd:
             driverobj._run_via_local_execution()
-            execute.assert_called_once_with(
+            run_shell_cmd.assert_called_once_with(
                 cmd="{x} >{x}.out 2>&1".format(x=driverobj._runscript_path),
                 cwd=driverobj.rundir,
                 log_output=True,
