+++ src/uwtools/tests/utils/test_processing.py
@@ -12,10 +12,10 @@ from uwtools.utils import processing
 def test_run_failure(caplog):
     processing.log.setLevel(logging.INFO)
     cmd = "expr 1 / 0"
-    success, output = processing.execute(cmd=cmd)
+    success, output = processing.run_shell_cmd(cmd=cmd)
     assert "division by zero" in output
     assert success is False
-    assert logged(caplog, "Executing: %s" % cmd)
+    assert logged(caplog, "Running: %s" % cmd)
     assert logged(caplog, "  Failed with status: 2")
     assert logged(caplog, "  Output:")
     assert logged(caplog, "    expr: division by zero")
@@ -24,9 +24,11 @@ def test_run_failure(caplog):
 def test_run_success(caplog, tmp_path):
     processing.log.setLevel(logging.INFO)
     cmd = "echo hello $FOO"
-    success, _ = processing.execute(cmd=cmd, cwd=tmp_path, env={"FOO": "bar"}, log_output=True)
+    success, _ = processing.run_shell_cmd(
+        cmd=cmd, cwd=tmp_path, env={"FOO": "bar"}, log_output=True
+    )
     assert success
-    assert logged(caplog, "Executing: %s" % cmd)
+    assert logged(caplog, "Running: %s" % cmd)
     assert logged(caplog, "  in %s" % tmp_path)
     assert logged(caplog, "  with environment variables:")
     assert logged(caplog, "    FOO=bar")
